Trying register as module for async
